Move functions to helpers.py
